% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_generated.r
\name{actor_get_preferences}
\alias{actor_get_preferences}
\title{This function is used for fetching the private settings or preferences associated with a user's account.}
\usage{
actor_get_preferences(token = NULL)
}
\value{
A list representing the account's preferences.
}
\description{
This function is used for fetching the private settings or preferences associated with a user's account.
}
\examples{
\dontrun{
account_preferences <- actor_get_preferences()
}
}
