% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_get_preferences}
\alias{app_bsky_actor_get_preferences}
\title{app_bsky_actor_get_preferences
Get private preferences attached to the account.}
\usage{
app_bsky_actor_get_preferences(.token = NULL, .return = c("json", "resp"))
}
\description{
app_bsky_actor_get_preferences
Get private preferences attached to the account.
}
