% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_get_profile}
\alias{app_bsky_actor_get_profile}
\title{app_bsky_actor_get_profile}
\usage{
app_bsky_actor_get_profile(actor, .token = NULL, .return = c("json", "resp"))
}
\description{
app_bsky_actor_get_profile
}
