% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_get_profiles}
\alias{app_bsky_actor_get_profiles}
\title{app_bsky_actor_get_profiles}
\usage{
app_bsky_actor_get_profiles(actors, .token = NULL, .return = c("json", "resp"))
}
\description{
app_bsky_actor_get_profiles
}
