% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_get_suggestions}
\alias{app_bsky_actor_get_suggestions}
\title{app_bsky_actor_get_suggestions
Get a list of actors suggested for following. Used in discovery UIs.}
\usage{
app_bsky_actor_get_suggestions(
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_actor_get_suggestions
Get a list of actors suggested for following. Used in discovery UIs.
}
