% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_put_preferences}
\alias{app_bsky_actor_put_preferences}
\title{app_bsky_actor_put_preferences
Sets the private preferences attached to the account.}
\usage{
app_bsky_actor_put_preferences(
  preferences,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_actor_put_preferences
Sets the private preferences attached to the account.
}
