% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_search_actors}
\alias{app_bsky_actor_search_actors}
\title{app_bsky_actor_search_actors
Find actors (profiles) matching search criteria.}
\usage{
app_bsky_actor_search_actors(
  term = NULL,
  q = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_actor_search_actors
Find actors (profiles) matching search criteria.
}
