% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_actor.R
\name{app_bsky_actor_search_actors_typeahead}
\alias{app_bsky_actor_search_actors_typeahead}
\title{app_bsky_actor_search_actors_typeahead
Find actor suggestions for a search term.}
\usage{
app_bsky_actor_search_actors_typeahead(
  term = NULL,
  q = NULL,
  limit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_actor_search_actors_typeahead
Find actor suggestions for a search term.
}
