% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_describe_feed_generator}
\alias{app_bsky_feed_describe_feed_generator}
\title{app_bsky_feed_describe_feed_generator
Returns information about a given feed generator including TOS & offered feed URIs}
\usage{
app_bsky_feed_describe_feed_generator(
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_describe_feed_generator
Returns information about a given feed generator including TOS & offered feed URIs
}
