% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_actor_feeds}
\alias{app_bsky_feed_get_actor_feeds}
\title{app_bsky_feed_get_actor_feeds
Retrieve a list of feeds created by a given actor}
\usage{
app_bsky_feed_get_actor_feeds(
  actor,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_actor_feeds
Retrieve a list of feeds created by a given actor
}
