% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_actor_likes}
\alias{app_bsky_feed_get_actor_likes}
\title{app_bsky_feed_get_actor_likes
A view of the posts liked by an actor.}
\usage{
app_bsky_feed_get_actor_likes(
  actor,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_actor_likes
A view of the posts liked by an actor.
}
