% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_author_feed}
\alias{app_bsky_feed_get_author_feed}
\title{app_bsky_feed_get_author_feed
A view of an actor's feed.}
\usage{
app_bsky_feed_get_author_feed(
  actor,
  limit = NULL,
  cursor = NULL,
  filter = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_author_feed
A view of an actor's feed.
}
