% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_feed_generator}
\alias{app_bsky_feed_get_feed_generator}
\title{app_bsky_feed_get_feed_generator
Get information about a specific feed offered by a feed generator, such as its online status}
\usage{
app_bsky_feed_get_feed_generator(
  feed,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_feed_generator
Get information about a specific feed offered by a feed generator, such as its online status
}
