% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_feed_generators}
\alias{app_bsky_feed_get_feed_generators}
\title{app_bsky_feed_get_feed_generators
Get information about a list of feed generators}
\usage{
app_bsky_feed_get_feed_generators(
  feeds,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_feed_generators
Get information about a list of feed generators
}
