% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_feed_skeleton}
\alias{app_bsky_feed_get_feed_skeleton}
\title{app_bsky_feed_get_feed_skeleton
A skeleton of a feed provided by a feed generator}
\usage{
app_bsky_feed_get_feed_skeleton(
  feed,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_feed_skeleton
A skeleton of a feed provided by a feed generator
}
