% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_list_feed}
\alias{app_bsky_feed_get_list_feed}
\title{app_bsky_feed_get_list_feed
A view of a recent posts from actors in a list}
\usage{
app_bsky_feed_get_list_feed(
  list,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_list_feed
A view of a recent posts from actors in a list
}
