% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_posts}
\alias{app_bsky_feed_get_posts}
\title{app_bsky_feed_get_posts
A view of an actor's feed.}
\usage{
app_bsky_feed_get_posts(uris, .token = NULL, .return = c("json", "resp"))
}
\description{
app_bsky_feed_get_posts
A view of an actor's feed.
}
