% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_reposted_by}
\alias{app_bsky_feed_get_reposted_by}
\title{app_bsky_feed_get_reposted_by}
\usage{
app_bsky_feed_get_reposted_by(
  uri,
  cid = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_reposted_by
}
