% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_suggested_feeds}
\alias{app_bsky_feed_get_suggested_feeds}
\title{app_bsky_feed_get_suggested_feeds
Get a list of suggested feeds for the viewer.}
\usage{
app_bsky_feed_get_suggested_feeds(
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_suggested_feeds
Get a list of suggested feeds for the viewer.
}
