% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_get_timeline}
\alias{app_bsky_feed_get_timeline}
\title{app_bsky_feed_get_timeline
A view of the user's home timeline.}
\usage{
app_bsky_feed_get_timeline(
  algorithm = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_get_timeline
A view of the user's home timeline.
}
