% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_feed.R
\name{app_bsky_feed_search_posts}
\alias{app_bsky_feed_search_posts}
\title{app_bsky_feed_search_posts
Find posts matching search criteria}
\usage{
app_bsky_feed_search_posts(
  q,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_feed_search_posts
Find posts matching search criteria
}
