% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_get_follows}
\alias{app_bsky_graph_get_follows}
\title{app_bsky_graph_get_follows
Who is an actor following?}
\usage{
app_bsky_graph_get_follows(
  actor,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_get_follows
Who is an actor following?
}
