% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_get_list}
\alias{app_bsky_graph_get_list}
\title{app_bsky_graph_get_list
Fetch a list of actors}
\usage{
app_bsky_graph_get_list(
  list,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_get_list
Fetch a list of actors
}
