% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_get_list_mutes}
\alias{app_bsky_graph_get_list_mutes}
\title{app_bsky_graph_get_list_mutes
Which lists is the requester's account muting?}
\usage{
app_bsky_graph_get_list_mutes(
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_get_list_mutes
Which lists is the requester's account muting?
}
