% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_get_lists}
\alias{app_bsky_graph_get_lists}
\title{app_bsky_graph_get_lists
Fetch a list of lists that belong to an actor}
\usage{
app_bsky_graph_get_lists(
  actor,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_get_lists
Fetch a list of lists that belong to an actor
}
