% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_get_suggested_follows_by_actor}
\alias{app_bsky_graph_get_suggested_follows_by_actor}
\title{app_bsky_graph_get_suggested_follows_by_actor
Get suggested follows related to a given actor.}
\usage{
app_bsky_graph_get_suggested_follows_by_actor(
  actor,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_get_suggested_follows_by_actor
Get suggested follows related to a given actor.
}
