% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_mute_actor_list}
\alias{app_bsky_graph_mute_actor_list}
\title{app_bsky_graph_mute_actor_list
Mute a list of actors.}
\usage{
app_bsky_graph_mute_actor_list(
  list,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_graph_mute_actor_list
Mute a list of actors.
}
