% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_graph.R
\name{app_bsky_graph_unmute_actor}
\alias{app_bsky_graph_unmute_actor}
\title{app_bsky_graph_unmute_actor
Unmute an actor by did or handle.}
\usage{
app_bsky_graph_unmute_actor(actor, .token = NULL, .return = c("json", "resp"))
}
\description{
app_bsky_graph_unmute_actor
Unmute an actor by did or handle.
}
