% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_notification.R
\name{app_bsky_notification_get_unread_count}
\alias{app_bsky_notification_get_unread_count}
\title{app_bsky_notification_get_unread_count}
\usage{
app_bsky_notification_get_unread_count(
  seenAt = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_notification_get_unread_count
}
