% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_notification.R
\name{app_bsky_notification_register_push}
\alias{app_bsky_notification_register_push}
\title{app_bsky_notification_register_push
Register for push notifications with a service}
\usage{
app_bsky_notification_register_push(
  serviceDid,
  token,
  platform,
  appId,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_notification_register_push
Register for push notifications with a service
}
