% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_notification.R
\name{app_bsky_notification_update_seen}
\alias{app_bsky_notification_update_seen}
\title{app_bsky_notification_update_seen
Notify server that the user has seen notifications.}
\usage{
app_bsky_notification_update_seen(
  seenAt,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_notification_update_seen
Notify server that the user has seen notifications.
}
