% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_unspecced.R
\name{app_bsky_unspecced_get_popular_feed_generators}
\alias{app_bsky_unspecced_get_popular_feed_generators}
\title{app_bsky_unspecced_get_popular_feed_generators
An unspecced view of globally popular feed generators}
\usage{
app_bsky_unspecced_get_popular_feed_generators(
  limit = NULL,
  cursor = NULL,
  query = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_unspecced_get_popular_feed_generators
An unspecced view of globally popular feed generators
}
