% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_unspecced.R
\name{app_bsky_unspecced_get_timeline_skeleton}
\alias{app_bsky_unspecced_get_timeline_skeleton}
\title{app_bsky_unspecced_get_timeline_skeleton
A skeleton of a timeline - UNSPECCED & WILL GO AWAY SOON}
\usage{
app_bsky_unspecced_get_timeline_skeleton(
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_unspecced_get_timeline_skeleton
A skeleton of a timeline - UNSPECCED & WILL GO AWAY SOON
}
