% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app_bsky_unspecced.R
\name{app_bsky_unspecced_search_actors_skeleton}
\alias{app_bsky_unspecced_search_actors_skeleton}
\title{app_bsky_unspecced_search_actors_skeleton
Backend Actors (profile) search, returning only skeleton}
\usage{
app_bsky_unspecced_search_actors_skeleton(
  q,
  typeahead = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
app_bsky_unspecced_search_actors_skeleton
Backend Actors (profile) search, returning only skeleton
}
