% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atr-package.R
\docType{package}
\name{atr-package}
\alias{atr}
\alias{atr-package}
\title{atr: What the Package Does (One Line, Title Case)}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The goal of atr is to wrap the AT Protocol (Authenticated Transfer Protocol)
behind Bluesky.
}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})

Authors:
\itemize{
  \item Benjamin Guinaudeau \email{wouter@vanatteveldt.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
}

}
\keyword{internal}
