% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atrrr-package.R
\docType{package}
\name{atrrr-package}
\alias{atrrr}
\alias{atrrr-package}
\title{atrrr: Wrapper for the 'AT' Protocol Behind 'Bluesky'}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The goal of atrrr is to wrap the AT Protocol (Authenticated Transfer Protocol)
behind Bluesky.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jbgruber.github.io/atrrr/}
  \item \url{https://github.com/JBGruber/atrrr}
  \item Report bugs at \url{https://github.com/JBGruber/atrrr/issues}
}

}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})

Authors:
\itemize{
  \item Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})
  \item Benjamin Guinaudeau \email{benjamin.guinaudeau@uni-konstanz.de} (\href{https://orcid.org/0000-0001-7206-6875}{ORCID}) [contributor]
  \item Fabio Votta \email{fabio.votta@gmail.com} (\href{https://orcid.org/0000-0002-3143-5942}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Nicholas Erskine \email{nicholas.erskine95@gmail.com} [contributor]
}

}
\keyword{internal}
