% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_disable_account_invites}
\alias{com_atproto_admin_disable_account_invites}
\title{com_atproto_admin_disable_account_invites
Disable an account from receiving new invite codes, but does not invalidate existing codes}
\usage{
com_atproto_admin_disable_account_invites(
  account,
  note = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_disable_account_invites
Disable an account from receiving new invite codes, but does not invalidate existing codes
}
