% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_disable_invite_codes}
\alias{com_atproto_admin_disable_invite_codes}
\title{com_atproto_admin_disable_invite_codes
Disable some set of codes and/or all codes associated with a set of users}
\usage{
com_atproto_admin_disable_invite_codes(
  codes = NULL,
  accounts = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_disable_invite_codes
Disable some set of codes and/or all codes associated with a set of users
}
