% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_invite_codes}
\alias{com_atproto_admin_get_invite_codes}
\title{com_atproto_admin_get_invite_codes
Admin view of invite codes}
\usage{
com_atproto_admin_get_invite_codes(
  sort = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_invite_codes
Admin view of invite codes
}
