% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_moderation_action}
\alias{com_atproto_admin_get_moderation_action}
\title{com_atproto_admin_get_moderation_action
View details about a moderation action.}
\usage{
com_atproto_admin_get_moderation_action(
  id,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_moderation_action
View details about a moderation action.
}
