% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_moderation_actions}
\alias{com_atproto_admin_get_moderation_actions}
\title{com_atproto_admin_get_moderation_actions
List moderation actions related to a subject.}
\usage{
com_atproto_admin_get_moderation_actions(
  subject = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_moderation_actions
List moderation actions related to a subject.
}
