% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_moderation_report}
\alias{com_atproto_admin_get_moderation_report}
\title{com_atproto_admin_get_moderation_report
View details about a moderation report.}
\usage{
com_atproto_admin_get_moderation_report(
  id,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_moderation_report
View details about a moderation report.
}
