% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_moderation_reports}
\alias{com_atproto_admin_get_moderation_reports}
\title{com_atproto_admin_get_moderation_reports
List moderation reports related to a subject.}
\usage{
com_atproto_admin_get_moderation_reports(
  subject = NULL,
  ignoreSubjects = NULL,
  actionedBy = NULL,
  reporters = NULL,
  resolved = NULL,
  actionType = NULL,
  limit = NULL,
  cursor = NULL,
  reverse = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_moderation_reports
List moderation reports related to a subject.
}
