% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_record}
\alias{com_atproto_admin_get_record}
\title{com_atproto_admin_get_record
View details about a record.}
\usage{
com_atproto_admin_get_record(
  uri,
  cid = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_get_record
View details about a record.
}
