% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_get_repo}
\alias{com_atproto_admin_get_repo}
\title{com_atproto_admin_get_repo
View details about a repository.}
\usage{
com_atproto_admin_get_repo(did, .token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_admin_get_repo
View details about a repository.
}
