% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_resolve_moderation_reports}
\alias{com_atproto_admin_resolve_moderation_reports}
\title{com_atproto_admin_resolve_moderation_reports
Resolve moderation reports by an action.}
\usage{
com_atproto_admin_resolve_moderation_reports(
  actionId,
  reportIds,
  createdBy,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_resolve_moderation_reports
Resolve moderation reports by an action.
}
