% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_search_repos}
\alias{com_atproto_admin_search_repos}
\title{com_atproto_admin_search_repos
Find repositories based on a search term.}
\usage{
com_atproto_admin_search_repos(
  term = NULL,
  q = NULL,
  invitedBy = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_search_repos
Find repositories based on a search term.
}
