% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_send_email}
\alias{com_atproto_admin_send_email}
\title{com_atproto_admin_send_email
Send email to a user's primary email address}
\usage{
com_atproto_admin_send_email(
  recipientDid,
  content,
  subject = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_send_email
Send email to a user's primary email address
}
