% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_take_moderation_action}
\alias{com_atproto_admin_take_moderation_action}
\title{com_atproto_admin_take_moderation_action
Take a moderation action on a repo.}
\usage{
com_atproto_admin_take_moderation_action(
  action,
  subject,
  reason,
  createdBy,
  subjectBlobCids = NULL,
  createLabelVals = NULL,
  negateLabelVals = NULL,
  durationInHours = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_take_moderation_action
Take a moderation action on a repo.
}
