% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_update_account_email}
\alias{com_atproto_admin_update_account_email}
\title{com_atproto_admin_update_account_email
Administrative action to update an account's email}
\usage{
com_atproto_admin_update_account_email(
  account,
  email,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_update_account_email
Administrative action to update an account's email
}
