% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_admin.R
\name{com_atproto_admin_update_account_handle}
\alias{com_atproto_admin_update_account_handle}
\title{com_atproto_admin_update_account_handle
Administrative action to update an account's handle}
\usage{
com_atproto_admin_update_account_handle(
  did,
  handle,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_admin_update_account_handle
Administrative action to update an account's handle
}
