% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_identity.R
\name{com_atproto_identity_resolve_handle}
\alias{com_atproto_identity_resolve_handle}
\title{com_atproto_identity_resolve_handle
Provides the DID of a repo.}
\usage{
com_atproto_identity_resolve_handle(
  handle,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_identity_resolve_handle
Provides the DID of a repo.
}
