% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_identity.R
\name{com_atproto_identity_update_handle}
\alias{com_atproto_identity_update_handle}
\title{com_atproto_identity_update_handle
Updates the handle of the account}
\usage{
com_atproto_identity_update_handle(
  handle,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_identity_update_handle
Updates the handle of the account
}
