% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_label.R
\name{com_atproto_label_query_labels}
\alias{com_atproto_label_query_labels}
\title{com_atproto_label_query_labels
Find labels relevant to the provided URI patterns.}
\usage{
com_atproto_label_query_labels(
  uriPatterns,
  sources = NULL,
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_label_query_labels
Find labels relevant to the provided URI patterns.
}
