% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_moderation.R
\name{com_atproto_moderation_create_report}
\alias{com_atproto_moderation_create_report}
\title{com_atproto_moderation_create_report
Report a repo or a record.}
\usage{
com_atproto_moderation_create_report(
  reasonType,
  subject,
  reason = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_moderation_create_report
Report a repo or a record.
}
