% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_apply_writes}
\alias{com_atproto_repo_apply_writes}
\title{com_atproto_repo_apply_writes
Apply a batch transaction of creates, updates, and deletes.}
\usage{
com_atproto_repo_apply_writes(
  repo,
  writes,
  validate = NULL,
  swapCommit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_apply_writes
Apply a batch transaction of creates, updates, and deletes.
}
