% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_create_record}
\alias{com_atproto_repo_create_record}
\title{com_atproto_repo_create_record
Create a new record.}
\usage{
com_atproto_repo_create_record(
  repo,
  collection,
  record,
  rkey = NULL,
  validate = NULL,
  swapCommit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_create_record
Create a new record.
}
