% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_delete_record}
\alias{com_atproto_repo_delete_record}
\title{com_atproto_repo_delete_record
Delete a record, or ensure it doesn't exist.}
\usage{
com_atproto_repo_delete_record(
  repo,
  collection,
  rkey,
  swapRecord = NULL,
  swapCommit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_delete_record
Delete a record, or ensure it doesn't exist.
}
