% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_describe_repo}
\alias{com_atproto_repo_describe_repo}
\title{com_atproto_repo_describe_repo
Get information about the repo, including the list of collections.}
\usage{
com_atproto_repo_describe_repo(
  repo,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_describe_repo
Get information about the repo, including the list of collections.
}
