% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_get_record}
\alias{com_atproto_repo_get_record}
\title{com_atproto_repo_get_record
Get a record.}
\usage{
com_atproto_repo_get_record(
  repo,
  collection,
  rkey,
  cid = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_get_record
Get a record.
}
