% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_list_records}
\alias{com_atproto_repo_list_records}
\title{com_atproto_repo_list_records
List a range of records in a collection.}
\usage{
com_atproto_repo_list_records(
  repo,
  collection,
  limit = NULL,
  cursor = NULL,
  rkeyStart = NULL,
  rkeyEnd = NULL,
  reverse = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_list_records
List a range of records in a collection.
}
