% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_repo.R
\name{com_atproto_repo_put_record}
\alias{com_atproto_repo_put_record}
\title{com_atproto_repo_put_record
Write a record, creating or updating it as needed.}
\usage{
com_atproto_repo_put_record(
  repo,
  collection,
  rkey,
  record,
  validate = NULL,
  swapRecord = NULL,
  swapCommit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_repo_put_record
Write a record, creating or updating it as needed.
}
