% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{com_atproto_repo_upload_blob2}
\alias{com_atproto_repo_upload_blob2}
\title{lexicon seems wrong. translated from https://atproto.com/blog/create-post#images-embeds}
\usage{
com_atproto_repo_upload_blob2(image, .token = NULL)
}
\description{
lexicon seems wrong. translated from https://atproto.com/blog/create-post#images-embeds
}
