% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_confirm_email}
\alias{com_atproto_server_confirm_email}
\title{com_atproto_server_confirm_email
Confirm an email using a token from com.atproto.server.requestEmailConfirmation.}
\usage{
com_atproto_server_confirm_email(
  email,
  token,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_confirm_email
Confirm an email using a token from com.atproto.server.requestEmailConfirmation.
}
