% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_create_account}
\alias{com_atproto_server_create_account}
\title{com_atproto_server_create_account
Create an account.}
\usage{
com_atproto_server_create_account(
  email,
  handle,
  password,
  did = NULL,
  inviteCode = NULL,
  recoveryKey = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_create_account
Create an account.
}
