% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_create_invite_code}
\alias{com_atproto_server_create_invite_code}
\title{com_atproto_server_create_invite_code
Create an invite code.}
\usage{
com_atproto_server_create_invite_code(
  useCount,
  forAccount = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_create_invite_code
Create an invite code.
}
