% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_create_invite_codes}
\alias{com_atproto_server_create_invite_codes}
\title{com_atproto_server_create_invite_codes
Create an invite code.}
\usage{
com_atproto_server_create_invite_codes(
  codeCount,
  useCount,
  forAccounts = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_create_invite_codes
Create an invite code.
}
