% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_delete_account}
\alias{com_atproto_server_delete_account}
\title{com_atproto_server_delete_account
Delete a user account with a token and password.}
\usage{
com_atproto_server_delete_account(
  did,
  password,
  token,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_delete_account
Delete a user account with a token and password.
}
