% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_describe_server}
\alias{com_atproto_server_describe_server}
\title{com_atproto_server_describe_server
Get a document describing the service's accounts configuration.}
\usage{
com_atproto_server_describe_server(.token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_server_describe_server
Get a document describing the service's accounts configuration.
}
