% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_list_app_passwords}
\alias{com_atproto_server_list_app_passwords}
\title{com_atproto_server_list_app_passwords
List all app-specific passwords.}
\usage{
com_atproto_server_list_app_passwords(
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_list_app_passwords
List all app-specific passwords.
}
