% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_refresh_session}
\alias{com_atproto_server_refresh_session}
\title{com_atproto_server_refresh_session
Refresh an authentication session.}
\usage{
com_atproto_server_refresh_session(.token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_server_refresh_session
Refresh an authentication session.
}
