% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_request_account_delete}
\alias{com_atproto_server_request_account_delete}
\title{com_atproto_server_request_account_delete
Initiate a user account deletion via email.}
\usage{
com_atproto_server_request_account_delete(
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_request_account_delete
Initiate a user account deletion via email.
}
