% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_request_email_confirmation}
\alias{com_atproto_server_request_email_confirmation}
\title{com_atproto_server_request_email_confirmation
Request an email with a code to confirm ownership of email}
\usage{
com_atproto_server_request_email_confirmation(
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_request_email_confirmation
Request an email with a code to confirm ownership of email
}
