% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_request_email_update}
\alias{com_atproto_server_request_email_update}
\title{com_atproto_server_request_email_update
Request a token in order to update email.}
\usage{
com_atproto_server_request_email_update(
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_request_email_update
Request a token in order to update email.
}
