% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_request_password_reset}
\alias{com_atproto_server_request_password_reset}
\title{com_atproto_server_request_password_reset
Initiate a user account password reset via email.}
\usage{
com_atproto_server_request_password_reset(
  email,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_request_password_reset
Initiate a user account password reset via email.
}
