% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_reset_password}
\alias{com_atproto_server_reset_password}
\title{com_atproto_server_reset_password
Reset a user account password using a token.}
\usage{
com_atproto_server_reset_password(
  token,
  password,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_reset_password
Reset a user account password using a token.
}
