% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_revoke_app_password}
\alias{com_atproto_server_revoke_app_password}
\title{com_atproto_server_revoke_app_password
Revoke an app-specific password by name.}
\usage{
com_atproto_server_revoke_app_password(
  name,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_revoke_app_password
Revoke an app-specific password by name.
}
