% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_server.R
\name{com_atproto_server_update_email}
\alias{com_atproto_server_update_email}
\title{com_atproto_server_update_email
Update an account's email.}
\usage{
com_atproto_server_update_email(
  email,
  token = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_server_update_email
Update an account's email.
}
