% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_blob}
\alias{com_atproto_sync_get_blob}
\title{com_atproto_sync_get_blob
Get a blob associated with a given repo.}
\usage{
com_atproto_sync_get_blob(did, cid, .token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_sync_get_blob
Get a blob associated with a given repo.
}
