% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_blocks}
\alias{com_atproto_sync_get_blocks}
\title{com_atproto_sync_get_blocks
Gets blocks from a given repo.}
\usage{
com_atproto_sync_get_blocks(
  did,
  cids,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_get_blocks
Gets blocks from a given repo.
}
