% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_checkout}
\alias{com_atproto_sync_get_checkout}
\title{com_atproto_sync_get_checkout
DEPRECATED - please use com.atproto.sync.getRepo instead}
\usage{
com_atproto_sync_get_checkout(did, .token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_sync_get_checkout
DEPRECATED - please use com.atproto.sync.getRepo instead
}
