% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_head}
\alias{com_atproto_sync_get_head}
\title{com_atproto_sync_get_head
DEPRECATED - please use com.atproto.sync.getLatestCommit instead}
\usage{
com_atproto_sync_get_head(did, .token = NULL, .return = c("json", "resp"))
}
\description{
com_atproto_sync_get_head
DEPRECATED - please use com.atproto.sync.getLatestCommit instead
}
