% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_latest_commit}
\alias{com_atproto_sync_get_latest_commit}
\title{com_atproto_sync_get_latest_commit
Gets the current commit CID & revision of the repo.}
\usage{
com_atproto_sync_get_latest_commit(
  did,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_get_latest_commit
Gets the current commit CID & revision of the repo.
}
