% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_record}
\alias{com_atproto_sync_get_record}
\title{com_atproto_sync_get_record
Gets blocks needed for existence or non-existence of record.}
\usage{
com_atproto_sync_get_record(
  did,
  collection,
  rkey,
  commit = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_get_record
Gets blocks needed for existence or non-existence of record.
}
