% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_get_repo}
\alias{com_atproto_sync_get_repo}
\title{com_atproto_sync_get_repo
Gets the did's repo, optionally catching up from a specific revision.}
\usage{
com_atproto_sync_get_repo(
  did,
  since = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_get_repo
Gets the did's repo, optionally catching up from a specific revision.
}
