% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_list_repos}
\alias{com_atproto_sync_list_repos}
\title{com_atproto_sync_list_repos
List dids and root cids of hosted repos}
\usage{
com_atproto_sync_list_repos(
  limit = NULL,
  cursor = NULL,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_list_repos
List dids and root cids of hosted repos
}
