% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_notify_of_update}
\alias{com_atproto_sync_notify_of_update}
\title{com_atproto_sync_notify_of_update
Notify a crawling service of a recent update. Often when a long break between updates causes the connection with the crawling service to break.}
\usage{
com_atproto_sync_notify_of_update(
  hostname,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_notify_of_update
Notify a crawling service of a recent update. Often when a long break between updates causes the connection with the crawling service to break.
}
