% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_com_atproto_sync.R
\name{com_atproto_sync_request_crawl}
\alias{com_atproto_sync_request_crawl}
\title{com_atproto_sync_request_crawl
Request a service to persistently crawl hosted repos.}
\usage{
com_atproto_sync_request_crawl(
  hostname,
  .token = NULL,
  .return = c("json", "resp")
)
}
\description{
com_atproto_sync_request_crawl
Request a service to persistently crawl hosted repos.
}
