% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{delete_skeet}
\alias{delete_skeet}
\title{Delete post}
\usage{
delete_skeet(post_url, .token = NULL)
}
\arguments{
\item{post_url}{URL of post to delete}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
Nothing. Called to delete a post.
}
\description{
Delete post
}
