% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{fetch_preview}
\alias{fetch_preview}
\title{Fetch link preview}
\usage{
fetch_preview(uri)
}
\arguments{
\item{uri}{URL or URI to fetch preview for.}
}
\value{
list strutured for use as preview_card
}
\description{
Fetch link preview
}
\examples{
\dontrun{
wiki_preview <- fetch_preview("https://en.wikipedia.org/wiki/AT_Protocol")
post_skeet("Do you know the AT Protocol?", preview_card = wiki_preview)
}
}
