% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{follow}
\alias{follow}
\alias{unfollow}
\title{Un/Follow an account}
\usage{
follow(actor, verbose = NULL, .token = NULL)

unfollow(actor, verbose = NULL, .token = NULL)
}
\arguments{
\item{actor}{User handle to follow or unfollow}

\item{verbose}{Whether to print status messages to the Console
(\code{TRUE}/\code{FALSE}). Package default (when \code{NULL}) is to have
status messages. Can be changed with \code{Sys.setenv(ATR_VERBOSE =
FALSE)}.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
list with URI and CID of the record (invisible).
}
\description{
Un/Follow an account
}
\details{
You can only unfollow accounts which you also followed through the
API/the package.
}
\examples{
\dontrun{
# follow our test account
follow("atpr.bsky.social")

# unfollow our test account
unfollow("atpr.bsky.social")
}
}
