% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_author_feed}
\alias{get_author_feed}
\title{A view of an actor's feed.}
\usage{
get_author_feed(actor, limit = 25L, cursor = NULL, .token = NULL)
}
\arguments{
\item{actor}{user handle to retrieve feed for.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results).}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of posts
}
\description{
A view of an actor's feed.
}
\examples{
feed <- get_author_feed("profmusgrave.bsky.social")
}
