% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_lexicon}
\alias{get_lexicon}
\title{Download new lexicon and save it into the inst folder}
\usage{
get_lexicon(path)
}
\arguments{
\item{path}{relative path from repository
\url{https://github.com/bluesky-social/atproto}.}
}
\value{
Nothing, downloads files into appropriate folder
}
\description{
Download new lexicon and save it into the inst folder
}
\examples{
get_lexicon("app/bsky/actor/getProfile.json")
}
