% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_replies}
\alias{get_replies}
\title{Get all replies}
\usage{
get_replies(post_url, .token = NULL)
}
\arguments{
\item{post_url}{the URL of a skeet.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame of skeets
}
\description{
Get all replies and replies on replies of a skeet.
}
\examples{
\dontrun{
get_replies("https://bsky.app/profile/jbgruber.bsky.social/post/3kbi57u4sys2l")
}
}
