% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_skeets_authored_by}
\alias{get_skeets_authored_by}
\title{A view of an actor's skeets.}
\usage{
get_skeets_authored_by(
  actor,
  limit = 25L,
  filter = NULL,
  cursor = NULL,
  parse = TRUE,
  verbose = NULL,
  .token = NULL
)
}
\arguments{
\item{actor}{user handle to retrieve feed for.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results). The function stops when the limit is reached, but you will
usually get a few items more than requested.}

\item{filter}{get only certain post/repost types. Possible values "posts_with_replies", "posts_no_replies", "posts_with_media", and "posts_and_author_threads".}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{verbose}{Whether to print status messages to the Console
(\code{TRUE}/\code{FALSE}). Package default (when \code{NULL}) is to have
status messages. Can be changed with \code{Sys.setenv(ATR_VERBOSE =
FALSE)}.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of posts
}
\description{
A view of an actor's skeets.
}
\examples{
\dontrun{
andrews_posts <- get_skeets_authored_by("andrew.heiss.phd")
}
}
