% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_thread}
\alias{get_thread}
\title{Get all skeets in a thread}
\usage{
get_thread(post_url, parse = TRUE, .token = NULL)
}
\arguments{
\item{post_url}{the URL of any skeet in a thread.}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame of skeets
}
\description{
Retrieve all skeets in a thread (all replies to an original skeet by any
author). It does not matter if you use the original skeet or any reply as
\code{post_url}.
}
\examples{
\dontrun{
get_thread("https://bsky.app/profile/jbgruber.bsky.social/post/3kbi57u4sys2l")
}
}
