% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actors.r
\name{get_user_info}
\alias{get_user_info}
\title{Query profile of an actor}
\usage{
get_user_info(actor, parse = TRUE, .token = NULL)
}
\arguments{
\item{actor}{user handle(s) to get information for.}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of found actors.
}
\description{
Query profile of an actor
}
\examples{
\dontrun{
rstats_user <- search_user("rstats", limit = 2L)
get_user_info(rstats_user$handle)
}
}
