% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{like_skeet}
\alias{like_skeet}
\alias{like_post}
\title{Like a skeet}
\usage{
like_skeet(post_url, verbose = NULL, .token = NULL)

like_post(post_url, verbose = NULL, .token = NULL)
}
\arguments{
\item{post_url}{URL or URI of post to delete.}

\item{verbose}{Whether to print status messages to the Console
(\code{TRUE}/\code{FALSE}). Package default (when \code{NULL}) is to have
status messages. Can be changed with \code{Sys.setenv(ATR_VERBOSE =
FALSE)}.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
invisible record information from the API
}
\description{
Like a skeet
}
\examples{
\dontrun{
# like a post
like_skeet("https://bsky.app/profile/jbgruber.bsky.social/post/3lcmymlgxwa2t")

# or feed in the result of some search
johannes_posts <- get_skeets_authored_by("jbgruber.bsky.social")
like_skeet(johannes_posts$uri)
}
}
