% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_lexicons}
\alias{list_lexicons}
\title{AT Protocol Lexicons}
\format{
\subsection{\code{list_lexicons}}{

A list with all lexicons available for the AT protocols.
\describe{
\item{names}{Name of the lexicon}
\item{values}{path relative to \url{https://github.com/bluesky-social/atproto/tree/main/lexicons}}
}
}
}
\source{
\url{https://github.com/bluesky-social/atproto}
}
\usage{
list_lexicons
}
\description{
Available lexicons for the AT Protocol (Authenticated Transfer Protocol)
}
\keyword{datasets}
