% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{print.bsky_token}
\alias{print.bsky_token}
\title{Print token}
\usage{
\method{print}{bsky_token}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bsky_token}}

\item{...}{not used.}
}
\value{
No return value, called to print the token to screen
}
\description{
Print a AT token
}
