% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicons.R
\name{read_lexicon}
\alias{read_lexicon}
\title{Download new lexicon and save it into the inst folder}
\usage{
read_lexicon(path)
}
\arguments{
\item{path}{relative path from repository
\url{https://github.com/bluesky-social/atproto}.}
}
\value{
The corresopnding json file
}
\description{
Download new lexicon and save it into the inst folder
}
\examples{
read_lexicon("app/bsky/actor/getProfile.json")
}
