% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regexs}
\alias{regexs}
\title{Regular expressions to match mentions and URLs}
\format{
\subsection{\code{regexs}}{

A list with regular expressions to match:
\describe{
\item{mention_regex}{Mentions of user handles like @atproto.com}
\item{url_regex}{URLs}
}
}
}
\source{
\url{https://atproto.com/blog/create-post#mentions-and-links}
}
\usage{
regexs
}
\description{
Regular expressions to match mentions and URLs
}
\keyword{datasets}
