% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actors.r
\name{search_actor}
\alias{search_actor}
\title{Find actors (profiles) matching search criteria.}
\usage{
search_actor(query, limit = 25L, cursor = NULL, parse = TRUE, .token = NULL)
}
\arguments{
\item{query}{The search query. Searches in user names and descriptions.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results).}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of found actors.
}
\description{
Find actors (profiles) matching search criteria.
}
\examples{
\dontrun{
search_actor("benguinaudeau.com")
search_actor("Blog: favstats.eu")
search_actor("JBGruber")
search_actor("@UvA_ASCoR")
search_actor("rstats", limit = 1000L)
}
}
