% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{auth}
\alias{auth}
\title{Authenticate for the  API}
\usage{
auth(
  user,
  password,
  domain = "https://bsky.app/",
  force_refresh = FALSE,
  verbose = TRUE,
  token = NULL
)
}
\arguments{
\item{user}{Your user handle (e.g, benguinaudeau.bsky.social).}

\item{password}{Your app password (usually created on
\url{https://bsky.app/settings/app-passwords}).}

\item{domain}{For now https://bsky.app/, but could change in the future.}

\item{force_refresh}{remove old token to request a fresh one.}

\item{verbose}{If TRUE, prints success message.}

\item{token}{(Stale) token object. Usually you don't need to use this. But if
you manage your own tokens and they get stale, you can use this parameter
and request a fresh token.}
}
\value{
An authentication token (invisible)
}
\description{
Authenticate for the  API
}
\details{
After requesting the token, it is safed in the location returned by
\code{file.path(tools::R_user_dir("atr", "cache"), Sys.getenv("BSKY_TOKEN", unset = "token.rds"))}. If you have multiple tokens, you can use
\code{Sys.setenv(BSKY_TOKEN = "filename.rds")} to save/load the token with a
different name.
}
