% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_feed}
\alias{get_feed}
\title{Get the skeets from a specific feed}
\usage{
get_feed(feed_url, limit = 25L, cursor = NULL, parse = TRUE, .token = NULL)
}
\arguments{
\item{feed_url}{The url of the requested feed}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results).}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of posts
}
\description{
Get the skeets that would be shown when you open the given feed
}
\examples{
\dontrun{
get_own_timeline()
get_own_timeline(algorithm = "reverse-chronological")
}
}
