% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{get_followers}
\alias{get_followers}
\alias{get_follows}
\title{Get followers and follows of an actor}
\usage{
get_followers(actor, limit = 25L, cursor = NULL, parse = TRUE, .token = NULL)

get_follows(actor, limit = 25L, cursor = NULL, parse = TRUE, .token = NULL)
}
\arguments{
\item{actor}{user handle to look up followers for.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results).}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of found actors.
}
\description{
Get followers and follows of an actor
}
\examples{
\dontrun{
get_followers("benguinaudeau.bsky.social")
get_follows("favstats.bsky.social")
}
}
