% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{get_followers}
\alias{get_followers}
\alias{get_follows}
\title{Get followers and follows of an actor}
\usage{
get_followers(
  actor,
  limit = 25L,
  cursor = NULL,
  parse = TRUE,
  verbose = NULL,
  .token = NULL
)

get_follows(
  actor,
  limit = 25L,
  cursor = NULL,
  parse = TRUE,
  verbose = NULL,
  .token = NULL
)
}
\arguments{
\item{actor}{user handle to look up followers for.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results). The function stops when the limit is reached, but you will
usually get a few items more than requested.}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{verbose}{Whether to print status messages to the Console
(\code{TRUE}/\code{FALSE}). Package default (when \code{NULL}) is to have
status messages. Can be changed with \code{Sys.setenv(ATR_VERBOSE =
  FALSE)}.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of found actors.
}
\description{
Get followers and follows of an actor
}
\examples{
\dontrun{
get_followers("benguinaudeau.bsky.social")

# get first page of results
follows_df <- get_follows("favstats.bsky.social", limit = 25L)

# continue same search, starting from the next match
follows_df2 <- get_follows("favstats.bsky.social", limit = 25L,
                           cursor = attr(follows_df, "last_cursor"))
}
}
