% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{get_skeets_authored_by}
\alias{get_skeets_authored_by}
\title{A view of an actor's feed.}
\usage{
get_skeets_authored_by(
  actor,
  limit = 25L,
  cursor = NULL,
  parse = TRUE,
  .token = NULL
)
}
\arguments{
\item{actor}{user handle to retrieve feed for.}

\item{limit}{Maximum number of records to return. For queries with more than
100 results, pagination is used automatically (one request per 100
results).}

\item{cursor}{Cursor for pagination (to pick up an old search).}

\item{parse}{Parse the results or return the original nested object sent by
the server.}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}
}
\value{
a data frame (or nested list) of posts
}
\description{
A view of an actor's feed.
}
\examples{
feed <- get_skeets_authored_by("profmusgrave.bsky.social")
}
