% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicons.R
\name{github_ls}
\alias{github_ls}
\title{List all files in github subfolders up to a certain depth}
\usage{
github_ls(repo, folder, max_depth = 10)
}
\arguments{
\item{repo}{target github repository}

\item{folder}{the subfolder that should be searched}

\item{depth}{the maximum depth taht should be explored}
}
\value{
The list of corresponding paths. If the depth is reached, it returns the folder where max-depth was reached
}
\description{
List all files in github subfolders up to a certain depth
}
\examples{
 github_ls(repo = "https://api.github.com/repos/bluesky-social/atproto/contents", folder = "lexicons")
}
