% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{post}
\alias{post}
\alias{post_skeet}
\alias{delete_skeet}
\alias{delete_post}
\title{Post a skeet}
\usage{
post(
  text,
  in_reply_to = NULL,
  image = NULL,
  image_alt = NULL,
  created_at = Sys.time(),
  .token = NULL
)

post_skeet(
  text,
  in_reply_to = NULL,
  image = NULL,
  image_alt = NULL,
  created_at = Sys.time(),
  .token = NULL
)

delete_skeet(post_url, .token = NULL)

delete_post(post_url, .token = NULL)
}
\arguments{
\item{text}{Text to post}

\item{in_reply_to}{URL or URI of a skeet this should reply to.}

\item{image}{path to an image to post}

\item{image_alt}{alt text for the image}

\item{created_at}{time stamp of the post}

\item{.token}{If you manage your own tokens, you can supply it here. Usually
\code{NULL} is OK and will automatically load or guide you to generate a
token.}

\item{post_url}{URL or URI of post to delete}
}
\value{
list of the URI and CID of the post (invisible)
}
\description{
Post a skeet
}
\examples{
\dontrun{
post("Hello from #rstats with {atr}")
}
}
